+++ R/psim_artif.R
@@ -11,9 +11,14 @@
 #' represent different integer values.
 #' @param start_date the start date of the temporal pattern.
 #' The date should be in the format `"yyyy-mm-dd"`.
-#' The GTP will normally cover a 1-year period.
+#' The 'gtp' will normally cover a 1-year period.
 #' @param poly (An sf or S4 object)
-#' a polygon shapefile defining the extent of the landscape
+#' a polygon shapefile defining the extent of the landscape.
+#' @param netw (An sf or S4 object)
+#' The network path of the landscape
+#' (e.g. road and/or street). Default: \code{NULL}.
+#' If provided each event is snapped to the closest
+#' network path/segment.
 #' @param n_origin number of locations to serve as
 #' origins for walkers. Default:\code{50}.
 #' @param restriction_feat (An S4 object) optional
@@ -61,16 +66,25 @@
 #' long-term trend. Options are:
 #' `"falling"`, `"stable"`,
 #' and `"rising"`. Default value is: `"stable"`.
+#' @param shortTerm type of short- to medium-term
+#' fluctuations (patterns) of the time series.
+#' Options are: \code{`"cyclical"` and `"acyclical"`}.
+#' Default is: \code{`"cyclical"`}.
+#' @param fPeak first seasonal
+#' peak of cyclical short term. Default value is \code{90}.
+#' Only used for `"cyclical"` short term pattern.
+#' @param s_band distance bandwidth within which
+#' the event re-occurences are maximized (i.e.,
+#' interactions are maximum). Specified as a vector of
+#' two distance values. Default: \code{c(0, 200)}.
+#' @param t_band temporal bandwidth within which
+#' event re-occurences are maximized (i.e., interactions
+#' are maximum). Specified as a vector of values (in days)
+#' \code{c(1, 5, 7, 14)}.
 #' @param slope slope of the long-term trend when
 #' an `"rising"` or `"falling"` trend is specified.
 #' Options: `"gentle"` or `"steep"`. The default value is
 #' set as \code{NULL} for the `stable` trend.
-#' @param first_pDate date of the first seasonal peak of
-#' the GTP (format: `"yyyy-mm-dd"`).
-#' Default value is \code{NULL}, in which first seasonal
-#' peak of 90 days is utilized.
-#' seasonal cycle of 180 days is utilized (that is,
-#' a seasonal cycle of 180 days).
 #' @param show.plot (logical) Shows GTP.
 #' Default is \code{FALSE}.
 #' @param show.data (TRUE or FALSE) To show the output
@@ -83,32 +97,38 @@
 #' @param ... additional arguments to pass from
 #' \code{gtp}, \code{walker} and \code{artif_spo}
 #' functions.
-#' @usage psim_artif(n_events=1000, start_date = "yyyy-mm-dd",
-#' poly, n_origin, restriction_feat=NULL, field,
+#' @usage psim_artif(n_events=1000, start_date = "2021-01-01",
+#' poly, netw = NULL, n_origin, restriction_feat=NULL, field,
 #' n_foci, foci_separation, mfocal = NULL, conc_type = "dispersed",
-#' p_ratio, s_threshold = 50, step_length = 20,
-#' trend = "stable", first_pDate=NULL,
+#' p_ratio=20, s_threshold = 50, step_length = 20,
+#' trend = "stable", shortTerm = "cyclical", fPeak=90,
+#' s_band = c(0, 200),
+#' t_band = c(1, 5, 10),
 #' slope = NULL, interactive = FALSE, show.plot=FALSE, show.data=FALSE, ...)
 #' @examples
 #' \dontrun{
 #'
 #' #load boundary and land use of Camden
-#' load(file = system.file("extdata", "camden.rda",
-#' package="stppSim"))
-#' boundary = camden$boundary # get boundary
-#' landuse = camden$landuse # get landuse
-#'
+#' #load(file = system.file("extdata", "camden.rda",
+#' #package="stppSim"))
+#' #boundary = camden$boundary # get boundary
+#' #landuse = camden$landuse # get landuse
+#' boundary <- stppSim:::boundary
+#' landuse <- stppSim:::landuse
 #' #In this example, we will use a minimal number of
 #' #'n_origin' (i.e. `20`) for faster computation:
 #'
 #' #simulate data
 #' simulated_stpp <- psim_artif(n_events=200, start_date = "2021-01-01",
-#' poly=boundary, n_origin=20, restriction_feat = NULL,
+#' poly=boundary, netw = NULL, n_origin=20, restriction_feat = NULL,
 #' field = NULL,
 #' n_foci=1, foci_separation = 10, mfocal = NULL,
 #' conc_type = "dispersed",
-#' p_ratio = 20, s_threshold = 50, step_length = 20,
-#' trend = "stable", first_pDate=NULL,
+#' p_ratio = 20, s_threshold = 50,
+#' step_length = 20,
+#' trend = "stable", shortTerm = "cyclical",
+#' fPeak=90, s_band = c(0, 200),
+#' t_band = c(1, 5, 10),
 #' slope = NULL, interactive = FALSE, show.plot=FALSE, show.data=FALSE)
 #'
 #' #If `n_events` is a vector of values,
@@ -126,26 +146,10 @@
 #' }
 #'
 #' @details
-#' Both the walkers
-#' and the landscape are configured arbitrarily (in accordance
-#' with the users knowledge of the domain.
-#' This function is computationally intensive. When run,
-#' an estimate of the expected computational time
-#' is first printed in the console for the user.
-#' Argument with the largest impacts on the computational
-#' time include `n_origin=50`, and `restriction_feat` when
-#' not \code{NULL}. Note: the `n_events`
-#' argument has little of no impacts on the
-#' computational time, and so it is recommended that
-#' that a user inputs a vector of several values
-#' to simulate.
-#' Lastly, in addition to exporting the simulated
-#' point patterns, the
-#' function also returns the simulated origins,
-#' the boundary and the restriction features
-#' (if supplied).
+#' Simulate artificial spatiotemporal patterns
+#' and interactions based user specifications.
 #' @return Returns a list of artificial spatiotemporal
-#' point patterns generated from scratch.
+#' point patterns based on user-defined parameters.
 #' @importFrom data.table rbindlist
 #' @importFrom SiMRiv resistanceFromShape
 #' @importFrom raster raster extent
@@ -157,25 +161,84 @@
 #' @importFrom graphics points legend
 #' @importFrom lubridate hms
 #' @importFrom ggplot2 ggplot
+#' @importFrom stats lm
+#' @importFrom otuSummary matrixConvert
+#' @importFrom sf st_nearest_points st_length
+#' st_cast
 #' @export
 #'
 
-psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
-                       poly, n_origin, restriction_feat=NULL, field=NA,
+psim_artif <- function(n_events=1000, start_date = "2021-01-01",
+                       poly, netw = NULL, n_origin, restriction_feat=NULL, field=NA,
                        n_foci, foci_separation, mfocal = NULL,
-                       conc_type = "dispersed", p_ratio,
-                       s_threshold = 50, step_length = 20,
-                       trend = "stable",
-                       first_pDate=NULL,
+                       conc_type = "dispersed", p_ratio=20,
+                       s_threshold = 50,
+                       step_length = 20,
+                       trend = "stable", shortTerm = "cyclical",
+                       fPeak=90, s_band = c(0, 200), t_band = c(1, 5, 10),
                        slope = NULL, interactive = FALSE, show.plot=FALSE, show.data=FALSE,...){
 
   #define global variables...
-  nrowh <- origins <- locid <- sn <- prob <- z <- NULL
+  nrowh <- origins <- locid <- sn <- prob <- z <-
+    datetime <- distVal <- ids<- filterField1 <-
+    filterField2 <- ID <- ID2 <- distVal1 <-
+    distVal2 <- rname <- cname <- st <- locID_sub <- NULL
 
   #first derive the spo object
   spo <- artif_spo(poly, n_origin =  n_origin, restriction_feat = restriction_feat,
                    n_foci=n_foci, foci_separation = foci_separation,
                    mfocal = mfocal, conc_type = conc_type, p_ratio = p_ratio)
+  spo
+  #
+  if(shortTerm == "acyclical" & is.null(s_band)) {
+    stop(" 's_band' argument cannot be NULL for 'acyclical' short term pattern!")
+  }
+
+  if(shortTerm == "acyclical" & is.null(t_band)) {
+    stop(" 't_band' argument cannot be NULL for 'acyclical' short term pattern!")
+  }
+
+  #if one value is inputted
+  if(length(s_band) != 2){
+    stop(" 's_band' needs to be vector of two values!")
+  }
+
+  if(shortTerm == "acyclical"){
+    #if one value is inputted
+    if(s_band[1] >= s_band[2]){
+      stop(" 's_band' value 2 cannot be greater than value 1!")
+    }
+
+    #check whether t_band is integer
+    check.integer <- function(N){
+      !grepl("[^[:digit:]]", format(N,  digits = 20, scientific = FALSE))
+    }
+
+
+    if(!check.integer(t_band)){
+      stop(" 't_band' value must be an integer!")
+    }
+
+    # if(t_band >= 90){
+    #   stop(" 't_band' value may be too large! Input a smaller value!")
+    # }
+
+  }
+
+  if(shortTerm == "cyclical"){
+    check.integer <- function(N){
+      !grepl("[^[:digit:]]", format(N,  digits = 20, scientific = FALSE))
+    }
+
+    if(!check.integer(fPeak)){
+      stop(" 'fPeak' value must be an integer!")
+    }
+
+    if(fPeak > 90){
+      stop(" 'fPeak' value must be less than or equal to 90!")
+    }
+
+  }
 
   #start_date <- as.Date(start_date)
 
@@ -185,8 +248,9 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   }
 
   #check first peak value
-  if(is.null(first_pDate)){
-    first_pDate <- as.Date(start_date) + 90
+  if(is.null(fPeak)){
+    #fPeak <- as.Date(start_date) + 90
+    fPeak <- 90
   }
 
   output <- list()
@@ -194,7 +258,7 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   start_date <- as.Date(start_date)
 
   #global variables
-    group_by <- idx <- . <- if_else <-
+  group_by <- idx <- . <- if_else <-
     tid <- NULL
 
   #define global variables
@@ -212,11 +276,26 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   #coordinates of the event origins (i.e. initial coordinates of the
   #simulation)
   coords <- spo$origins %>%
-    select(x, y)
+    dplyr::select(x, y)
+
+
+  #testing if crs' are the same
+  if(!is.null(netw)){
+    crs_poly <- sf::st_crs(poly)$epsg
+    crs_netw <- sf::st_crs(netw)$epsg
+
+    if(crs_poly != crs_netw){
+      stop("Project of 'poly' and 'netw' shapefiles are not the same!! ")
+    }
+  }
 
   #simulate the global temporal pattern
-  gtp <- gtp(start_date = start_date, trend, slope=slope, first_pDate=first_pDate,
+  gtp <- gtp(start_date = start_date, trend = trend, slope=slope,
+             shortTerm = shortTerm,
+             fPeak=fPeak,
              show.plot=show.plot) #"01-01"
+
+  #gtp$data <- rep(20, 366)
 
 
   #prepare date
@@ -224,7 +303,7 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   t <- seq(0, 365, by = 1)
   t2 <- t1 + t #list of dates
 
-   #test polygon geometry
+  #test polygon geometry
   if(!is.null(poly)){
     #-----
     poly_tester(poly)
@@ -237,51 +316,50 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   spo_xy <- as_tibble(spo$origins) %>%
     dplyr::select(x, y)
 
-
-  if(interactive == TRUE){
-  #Create spatial and temporal models
-  #-----
-
-  cat(paste("#-------------------------------------------#",
-              "#-------------------------------------------#",
-              "#-------------------------------------------#",sep="\n"))
-  cat("                                             ")
-
-  query1 <- readline(prompt = "Preview Spatial and Temporal model? (Y / N):")
-
-  if(query1 %in% c("Y", "y")){
-
-    stm(pt = spo$origins %>%
-                    select(x, y, prob), poly=spo$poly, df=gtp$data,
-                    crsys = projection(spo$poly), display_output = TRUE)
-    #flush.console()
-    cat(paste("#-----------------#",
-              "#-----------------#",
-              "#-----------------#",sep="\n"))
-    cat("                   ")
-
-    query2 <- readline(prompt = "Continue? (Y / N):")
-
-    if(!query2 %in% c("N", "n", "Y", "y")){
-      stop("Invalid input! 'Y' or 'N', expected! Process terminated!")
-    }
-
-    if(query2 %in% c("N", "n")){
-      stop("Process terminated!")
-    }
-
-    if(query2 %in% c("Y", "y")){
-      #continue processing
-    }
+  #check the length and
+  #and the possible outputs
+  if(length(n_events) > 1 & shortTerm == "acyclical"){
+    cat(paste0("Note: One event set (i.e. n_events=", n_events[1], ")",
+               " will be generated for acyclical short term pattern!!"))
   }
 
-  # if(!query1 %in% c("N","n")){
-  #     stop("Invalid input! 'N' or 'n' expected! Process terminated!")
-  #   }
-  #
-  # if(query1 %in% c("N","n")){
-  #   #do nothing
-  # }
+
+  if(interactive == TRUE){
+    #Create spatial and temporal models
+    #-----
+
+    cat(paste("#-------------------------------------------#",
+              "#-------------------------------------------#",
+              "#-------------------------------------------#",sep="\n"))
+    cat("                                             ")
+
+    query1 <- readline(prompt = "Preview Spatial and Temporal model? (Y / N):")
+
+    if(query1 %in% c("Y", "y")){
+
+      stm(pt = spo$origins %>%
+            select(x, y, prob), poly=spo$poly, df=gtp$data,
+          crsys = projection(spo$poly), display_output = TRUE)
+      #flush.console()
+      cat(paste("#-----------------#",
+                "#-----------------#",
+                "#-----------------#",sep="\n"))
+      cat("                   ")
+
+      query2 <- readline(prompt = "Continue? (Y / N):")
+
+      if(!query2 %in% c("N", "n", "Y", "y")){
+        stop("Invalid input! 'Y' or 'N', expected! Process terminated!")
+      }
+
+      if(query2 %in% c("N", "n")){
+        stop("Process terminated!")
+      }
+
+      if(query2 %in% c("Y", "y")){
+        #continue processing
+      }
+    }
 
   }
 
@@ -290,7 +368,7 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   options(digits.secs = 5)
   tme1 <- Sys.time()
   event_loc_N <- lapply(n, function(n)
-    stppSim::walker(n, s_threshold = s_threshold,
+    stppSim::walker(n, s_threshold = s_threshold, #
                     poly=poly, restriction_feat = restriction_feat,
                     field = field,
                     coords=as.vector(unlist(spo_xy[1,],)),
@@ -304,8 +382,17 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   flush.console()
   time_elapse <- time_elapse + (time_elapse * 0.1)#add 10%
   cat("#=====")
-  cat("The expected computational time for the process is:",paste(time_elapse, " minutes", sep=""),sep=" ")
+  cat("*--------- Expected time of execution: ",paste(time_elapse, " minutes ---------*", sep=""),sep=" ")
   cat("=====#")
+
+  #preview
+  #-----------------
+  loc_N <- rbindlist(event_loc_N,
+                     use.names=TRUE, fill=TRUE, idcol="tid")
+  loc_N <- loc_N%>%dplyr::filter(tid==1)
+  #dev.new()
+  plot(loc_N$x, loc_N$y)
+  #-----------------
 
   #the actual process
   stp_All <- NULL
@@ -321,36 +408,254 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
     )
 
     loc_N <- rbindlist(event_loc_N,
-                          use.names=TRUE, fill=TRUE, idcol="tid")
+                       use.names=TRUE, fill=TRUE, idcol="tid")
 
     loc_N <- loc_N %>%
       mutate(locid=b, prob=spo$origins$prob[b]) %>%
       mutate(time=format(((tid-1) + as.Date(start_date) + hms(time)),
                          "%Y-%m-%d %H:%M:%S"))%>%
-      rename(datetime=time)
+      dplyr::rename(datetime=time)
 
     stp_All <- stp_All %>%
       bind_rows(loc_N)
-    }
 
-  #generate all the results
-  for(h in seq_len(length(n_events))){
+  }
 
+  #This is the second option
+  #-----------------------
+  if(shortTerm == "cyclical"){
     #add idx
-    stp_All_ <- stp_All %>%
+    stp_All_bk <- stp_All %>%
       rownames_to_column('ID') #%>% #add row as column
 
-    #sample to derive required number
-    samp_idx <- as.numeric(sample(stp_All_$ID, size = n_events[h],
-                                replace = FALSE, prob = stp_All_$prob)) #%>
+    samp_idx <- as.numeric(sample(stp_All_bk$ID,
+                                  size = round(0.05 * nrow(stp_All_bk), digits=0),
+                                  replace = FALSE)) #%>
+    #select 5%
+    stp_All_ <- stp_All_bk[samp_idx, ]
 
-    stp_All_ <- stp_All_[samp_idx, ]
+    #regenerate IDs
+    stp_All_ <-  stp_All_ %>%
+      dplyr::select(-c(ID))%>%
+      rownames_to_column('ID') #%
+  }
+
+  #-----------------------
+  if(shortTerm == "acyclical"){
+
+    #to adjust the baseline of time series
+    datxy <- stp_All
+
+    #divide the data and keep backup
+    s_id <- sample(1:nrow(datxy), round(nrow(datxy)*0.75, digits = 0), replace=FALSE)
+    div_75 <- datxy[s_id, ]
+
+    div_25 <- datxy[!1:nrow(datxy)%in%s_id,]
+
+    datxy_plot <- div_75 %>%
+      #datxy_plot <- stp_All_sub %>%
+      dplyr::mutate(t = as.Date(substr(datetime, 1,10)))%>%
+      group_by(t) %>%
+      dplyr::summarise(n = dplyr::n()) %>%
+      mutate(time = as.numeric(difftime(t, as.Date("2020-12-31"), units="days")))%>%
+      as.data.frame()
+
+    time <- data.frame(time=1:365)
+
+    datxy_plot <- time %>%
+      left_join(datxy_plot) %>%
+      dplyr::mutate(n = replace_na(n, 0))
+
+    #fit and plot
+    loessData1 <- data.frame(
+      x = 1:nrow(datxy_plot),
+      y = predict(lm(n~time, datxy_plot)),
+      method = "loess()"
+    )
+
+    loessData1 <- round(loessData1$y, digits = 0)
+
+    #randomly remove point for each day data
+    filtered_stp_All <- NULL
+
+    for(rmv in seq_len(length(loessData1))){ #rmv = 1
+
+      #data to either reduce or increase
+      stp_All_on_day <- div_75 %>%
+        dplyr::filter(tid == rmv)
+
+      if(datxy_plot$n[rmv] >= loessData1[rmv]){
+        orig_pt <- sample(1:datxy_plot$n[rmv], loessData1[rmv], replace = FALSE)
+        todaysData <- stp_All_on_day[orig_pt, ]
+      }
+
+      if(datxy_plot$n[rmv] < loessData1[rmv]){
+        #first borrow the remainder
+        rem_D <- div_25[sample(1:nrow(div_25), (loessData1[rmv] - datxy_plot$n[rmv]), replace = F),]
+        todaysData <- rbind(stp_All_on_day, rem_D)
+      }
+
+      filtered_stp_All <- rbind(filtered_stp_All, todaysData)
+
+    }
+
+    #-----------------------------------------------
+    #integrate the spatiotemporal sign.
+    #-----------------------------------------------
+    event_Collate <- NULL
+
+    fN_final_dt_convert <- NULL
+
+    #loop by origin
+    ori_sn <- unique(filtered_stp_All$locid)[order(unique(filtered_stp_All$locid))]
+
+    init_n <- 0
+
+    for(or in seq_len(length(ori_sn))){ #or=1
+      ##while(init_n <= n_events * 2) {
+
+      sub_Dat <- filtered_stp_All %>%
+        rownames_to_column('locID_sub') %>%
+        dplyr::filter(locid == ori_sn[or])
+
+      sample_sub_Dat <- sub_Dat[sample(1:nrow(sub_Dat),
+                                       round(nrow(sub_Dat)*0.5, digits = 0),
+                                       replace = FALSE),]
+
+      tme <-as.numeric(as.Date(sample_sub_Dat$datetime))#[1:10]
+      dt = dist(tme)
+
+      #for a specified time threshold
+      dt_convert <- matrixConvert(dt, colname = c("cname", "rname", "distVal"))
+
+      #get t threshold
+      t_st_List <- t_band
+
+      #maximize the occurence of t threshold
+      dt_conver_Wthres <- dt_convert %>%
+        tibble()%>%
+        dplyr::filter(distVal %in%  t_st_List) %>% #[1]
+        dplyr::rename(distVal1 = distVal)
+
+      #apply distance threshold
+      xy <- data.frame(x=sample_sub_Dat$x, y=sample_sub_Dat$y)#[1:10,]
+
+      ds <- dist(xy)
+
+      ds_convert <- matrixConvert(ds, colname = c("cname", "rname", "distVal"))
+
+      #get s threshold
+      nm <- s_band
+
+      #maximize the occurence of s threshold
+      ds_convert2 <- ds_convert %>%
+        dplyr::rename(distVal2 = distVal) %>%
+        dplyr::mutate(distVal2 = round(distVal2, digits = 0)) %>%
+        dplyr::filter(distVal2 %in% c(nm[1]:nm[2]))
+
+      f_count <- ds_convert2 %>%
+        dplyr::left_join(dt_conver_Wthres) %>%
+        dplyr::filter(!is.na(distVal1))
+
+      f_count_col <- f_count %>%
+        dplyr::group_by(cname)%>%
+        dplyr::count()%>%
+        dplyr::arrange(desc(n))%>%
+        data.frame()%>%
+        dplyr::top_frac(0.05)
+
+      f_count_row <- f_count %>%
+        dplyr::group_by(rname)%>%
+        dplyr::count()%>%
+        dplyr::arrange(desc(n))%>%
+        data.frame()%>%
+        dplyr::top_frac(0.05)
+
+
+      dt_conver_Wthres_Comb <- data.frame(ids = c(f_count_row$rname, f_count_col$cname))
+      ids <- unique(dt_conver_Wthres_Comb$ids)
+      sample_sub_DatNew <- sample_sub_Dat[as.numeric(ids),]
+
+      #-----------------------------------------------------------
+      if(length(ids) != 0){
+        event_Collate <- rbind(event_Collate,  sample_sub_DatNew)
+      }
+      #----------------event_Collate--------------------------...#
+
+      flush.console()
+      print(st)
+      print(nrow(event_Collate))
+
+      event_Collate <- event_Collate %>%
+        dplyr::filter(!duplicated(locID_sub))
+
+      init_n <- nrow(event_Collate)
+
+      flush.console()
+      print(or)
+      print(init_n)
+
+    }
+  }
+
+  #------------------------------------------.......#
+  stp_All_bk <- event_Collate
+  #------------------------------------------.......#
+
+  #finalizing!
+  #--------------------------------------#stp_All
+
+  #generate all the results
+  for(h in seq_len(length(n_events))){ #h<-1
+
+    #Also if the final list is very small compare
+    #to the list needed
+    if(nrow(stp_All_bk) < round(n_events[h]*1.5, digits = 0)){
+      cat(paste0("*------------| A total of ",  round(nrow(stp_All_bk)*0.75, digits = 0),
+                 " events are generated! |--------------*"))
+
+      n_events_h <- round(nrow(stp_All_bk)*0.75, digits = 0)
+
+      #sample to derive required number
+      samp_idx <- as.numeric(sample(1:nrow(stp_All_bk), size = n_events_h,
+                                    replace = FALSE, prob = stp_All_bk$prob)) #%>
+    }
+
+    if(nrow(stp_All_bk) >= round(n_events[h]*1.5, digits = 0)){
+      #sample to derive required number
+      samp_idx <- as.numeric(sample(1:nrow(stp_All_bk), size = n_events[h],
+                                    replace = FALSE, prob = stp_All_bk$prob)) #%>
+
+    }
+
+    stp_All_fn <- stp_All_bk[samp_idx, ]
 
     #sort
-    stp_All_ <- stp_All_ %>%
-      arrange(locid, tid, sn)
+    stp_All_fn <- stp_All_fn %>%
+      arrange(locid, tid, sn) #%>%
 
-    output[h] <- list(stp_All_)
+    output[h] <- list(stp_All_fn)
+  }
+
+  #if network path is provided
+  if(!is.null(netw)){ #netw <- netw_
+
+    flush.console()
+    print("***------Generating network data, processing....:")
+
+    for(g in seq_len(length(n_events))){ #g<-1
+
+      #convert point to geometry type
+      output_pt <- data.frame(output[g]) %>%
+        st_as_sf(coords = c("x", "y"), crs = crs_netw, remove =F)#%>%
+
+      snappedData <- snap_points_to_lines(points=output_pt,
+                                          lines=netw,
+                                          verbose = FALSE)
+
+      output[[g]]$x <- st_coordinates(snappedData)[,1]
+      output[[g]]$y <- st_coordinates(snappedData)[,2]
+    }
   }
 
   #add the origins
@@ -359,5 +664,12 @@ psim_artif <- function(n_events=1000, start_date = "yyyy-mm-dd",
   output$poly <- spo$poly
   output$resist <- restriction_feat
 
+  #insert network path
+  if(!is.null(netw)){
+    output$netw <- netw
+  }
+
   return(output)
+
+
 }