+++ NEWS.md
@@ -1,3 +1,19 @@
+# stppSim 1.3.0
+
+## Bug Fixing:
+
+- Resolved the issue of dependencies on retiring package (`rgdal`): 'https://github.com/MAnalytics/stppSim/issues/2'
+
+## New changes:
+
+- Introduced two new parameters, `s_band` and `t_band`, to the `psim_artif` function.
+- Developed a new function named `snap_points_to_line`, which has been incorporated into both psim_artif and psim_real functions.
+- Added a new function named `NRepeat` for assessing space-time point interaction
+- Incorporated a new argument, s_range, to `psim_real` function to establish the maximum spatial range.
+- Introduced the `s_interaction` argument to `psim_real` function to determine the spatial properties of point interactions.
+- Added the `tolerance` parameter, allowing users to set a p-value threshold for the detection of point interaction.
+
+
 # stppSim 1.2.7
 
 ## Bug Fixing